requirements.txt
# Core dependencies for CPU-based RAG system
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0

# LLM support (CPU)
llama-cpp-python==0.2.90

# RAG and embeddings
faiss-cpu==1.8.0
sentence-transformers==3.3.1
transformers==4.48.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
pandas>=2.0.0
openpyxl==3.1.5

# Flask web framework
Flask==3.0.3
Flask-CORS==4.0.1

# Environment and utilities
python-dotenv==1.0.1
SQLAlchemy==2.0.32

# Additional utilities
numpy>=1.24.0
torch>=2.0.0
tqdm==4.66.5
